+++ Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/Jobs/Collision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 874e448619fc4d54f900f52f83708a6c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
