+++ Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/Jobs/Particle.cs
@@ -1,46 +1,12 @@
-using System.Collections;
-using System.Collections.Generic;
 using UniGLTF.SpringBoneJobs.Blittables;
 using Unity.Collections;
 using Unity.Jobs;
 using UnityEngine;
 using UnityEngine.Jobs;
 
+
 namespace RotateParticle.Jobs
 {
-    public enum TransformType
-    {
-        Center,
-        WarpRootParent,
-        WarpRoot,
-        Particle,
-    }
-
-    public static class TransformTypeExtensions
-    {
-        public static bool PositionInput(this TransformType t)
-        {
-            return t == TransformType.WarpRoot;
-        }
-
-        public static bool Movable(this TransformType t)
-        {
-            return t == TransformType.Particle;
-        }
-
-        public static bool Writable(this TransformType t)
-        {
-            switch (t)
-            {
-                case TransformType.WarpRoot:
-                case TransformType.Particle:
-                    return true;
-                default:
-                    return false;
-            }
-        }
-    }
-
     public struct TransformInfo
     {
         public TransformType TransformType;
@@ -69,23 +35,7 @@ namespace RotateParticle.Jobs
         }
     }
 
-    public struct WarpInfo
-    {
-        public int StartIndex;
-        public int EndIndex;
-    }
-
     // [Input]
-    public struct InputColliderJob : IJobParallelForTransform
-    {
-        [WriteOnly] public NativeArray<Matrix4x4> CurrentCollider;
-
-        public void Execute(int index, TransformAccess transform)
-        {
-            CurrentCollider[index] = transform.localToWorldMatrix;
-        }
-    }
-
     public struct InputTransformJob : IJobParallelForTransform
     {
         [ReadOnly] public NativeArray<TransformInfo> Info;
@@ -145,39 +95,6 @@ namespace RotateParticle.Jobs
             }
 
             NextRotations[index] = CurrentTransforms[index].Rotation;
-        }
-    }
-
-    public struct CollisionJob : IJobParallelFor
-    {
-        [ReadOnly] public NativeArray<BlittableCollider> Colliders;
-        [ReadOnly] public NativeArray<Matrix4x4> CurrentColliders;
-        [WriteOnly] public NativeArray<Vector3> NextPositions;
-
-        public void Execute(int index)
-        {
-            for (int i = 0; i < Colliders.Length; ++i)
-            {
-                var c = Colliders[i];
-                switch (c.colliderType)
-                {
-                    case BlittableColliderType.Sphere:
-                        break;
-
-                    default:
-                        // TODO
-                        break;
-                }
-            }
-            // var d = Vector3.Distance(from, to);
-            // if (d > (ra + rb))
-            // {
-            //     resolved = default;
-            //     return false;
-            // }
-            // Vector3 normal = (to - from).normalized;
-            // resolved = new(from, from + normal * (d - rb));
-            // return true;
         }
     }
