+++ Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/Jobs/Particle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 69453c7639ad8df4c94cbe3dae50e72c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
