+++ Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/Jobs/TransformType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 56860caf8e0babc41807521ed55a4f20
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
