+++ Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/Jobs/Warp.cs
@@ -1,0 +1,8 @@
+namespace RotateParticle.Jobs
+{
+    public struct WarpInfo
+    {
+        public int StartIndex;
+        public int EndIndex;
+    }
+}
